\name{Data_Fish_rec.dta}
\docType{data}
\alias{Data_Fish_rec}
\title{Données de recrutement de l'Anchois en Méditerranée et variables environemmentales associées}
\description{
  Tableau de collecte des données biologiques individuelles
}
\usage{Data_Fish_rec.dta}
\format{Une dataframe}
\arguments{
\item{Recrue}{Nombre de recrue}

\item{Biomasse feconde}{Biomasse de reproducteurs}

\item{SST}{Température eaux de surface}

\item{Phytoplancton}{Concentration phytoplancton}

\item{Zooplancton}{Concentration zooplancton}

\item{WeMOi}{Indice climatique Ouest Mediterranée}
}
\source{P. Brosset}
\keyword{datasets}
