\name{Fish_estuaries.dta}
\docType{data}
\alias{Data_Fish_est}
\title{Données de distibution et de condition corporelle de 3 poissons estuariens et variables environemmentales associées}
\description{
  Tableau de collecte des données biologiques individuelles
}
\usage{Fish_estuaries.dta}
\format{Une dataframe}
\arguments{
\item{Station}{Station de collecte}

\item{Nb_sole}{Densité de sole par unité de surface}

\item{Nb_bar}{Densité de bar par unité de surface}

\item{Nb_flet}{Densité de flet par unité de surface}

\item{K_sole}{Condition corporelle de la sole}

\item{K_bar}{Condition corporelle du bar}

\item{K_flet}{Condition corporelle du flet}

\item{Nut}{Concentration nutriments}

\item{Sal}{Salinité}

\item{SST}{Température eaux de surface}

\item{Chla}{Concentration chlorophylle a}

}
\source{P. Brosset}
\keyword{datasets}
