\name{logbook.dta}
\docType{data}
\alias{logbook.dta}
\title{Données logbook Mauritaniennes pour 2 les Sardines et les Sardinelles}
\description{
  Extrait des déclaration des captures collectées par la DSPCM. C'est déclarations sont obbligatoires pour la pêche industrielle opérant dans la ZEE Mauritanienne
}
\usage{logbook.dta}
\format{Une dataframe}
\arguments{
\item{an}{Année de la capture}

\item{mois}{Mois de la capture}

\item{secteur}{Rectangle statistique de la capture - référentiel logbook.zr4n}

\item{codesp}{Espèce déclarée - referentiel logbook.zr2n}

\item{portfin}{Port de Débarquement}

\item{typelic}{Type de licence utilisé par les navires }

\item{codeng}{Code de l'engin utilisé}

\item{nation}{Pavillon des bateaux}

\item{typnav}{Type de bavire (congélateurs - )}

\item{classe_long}{Classe de longueur des bateaux}

\item{classe_tjb}{Classe de jauge brute des bateaux}

\item{classe_kw}{Classe de puissance en Kw des bateaux}

\item{classe_calpoids}{Classe de contenance des cales des bateaux - pas sur à 100 }

\item{dec_cons}{Decennie de construction des bateaux}

\item{dec_hom}{Dizaines d'hommes à bord}

\item{poids}{Poids de la capture}


}
\source{DSPCM Logbooks}
\keyword{datasets}
