\name{logbook.zr2n}
\docType{data}
\alias{logbook.zr2n}
\title{Référentiel des espèces}
\description{
  Contient le tableau de correspondance entre les codes especes utilisés dans les déclarations et les noms d'espèces correspondant
}
\usage{logbook.zr2n}
\format{Une dataframe}
\arguments{

\item{codesp}{Espèce déclarée}

\item{codfao}{Codage FAO de l'espèce}

\item{nomespco}{Nom commercial de l'espèce}

\item{contenu}{Nom taxonomique de l'espèce}

}
\source{DSPCM Logbooks}
\keyword{datasets}
