\name{logbook.zr4n}
\docType{data}
\alias{logbook.zr4n}
\title{Référentiel des rectangles statistiques}
\description{
  Contient le référentiel des rectangles statistiques, y compris le dessin de ceux-ci, le polygone
}
\usage{logbook.zr4n}
\format{Une spatial dataframe / sf object}
\arguments{

\item{secteur}{Code du secteur}

\item{geom}{Object spatial / Polygone du rectangle}

}
\source{DSPCM Logbooks}
\keyword{datasets}
