\name{maturite.dta}
\docType{data}
\alias{maturite.dta}
\title{Données de taille et de maturité collectées lors du projet européen DEMERSTEM dans le cadre de l'identification des stocks}
\description{
  Tableau de collecte des données biologiques individuelles
}
\usage{maturite.dta}
\format{Une dataframe}
\arguments{
\item{pays}{MRT ou SEN pour Mauritanie / Sénégal}

\item{Nom}{Nom du port de collecte de la données}

\item{annee}{Nom du port de collecte de la données}

\item{mois}{Nom du port de collecte de la données}

\item{code_fao}{Code FAO de l'espece : GPW / Thiof / Epinephelus Aeneus}

\item{lt}{Longueur Totale (mm) }

\item{maturity}{Maturité DEMERSTEM- Protocoles d'échantillonnage biologique
1 - IMMATURE/REPOS
2 - DEVELOPPEMENT/EN VOIE DE MATURATION
3 - PRESQUE À MATURITÉ
4 - GRAVIDES (FEMELLES) OU MATURES (MÂLES)
5 - VIDE (POST-PONTE)
}

\item{code_sexe}{1- Masculin 2- Feminin 3- Indéterminé}


}
\source{DEMERSTEM EU project}
\keyword{datasets}
