\name{taille_poids.dta}
\docType{data}
\alias{taille_poids}
\title{Données de taille et de poids individuels collectées lors du projet européen DEMERSTEM dans le cadre de l'identification des stocks}
\description{
  Tableau de collecte des données biologiques individuelles
}
\usage{taille_poids.dta}
\format{Une dataframe}
\arguments{
\item{pays}{MRT ou SEN pour Mauritanie / Sénégal}

\item{Nom}{Nom du port de collecte de la données}

\item{annee}{Nom du port de collecte de la données}

\item{mois}{Nom du port de collecte de la données}

\item{code_fao}{Code FAO de l'espece : GPW / Thiof / Epinephelus Aeneus}

\item{lt}{Longueur Totale (mm) }

\item{pt}{poids total (g) }

\item{code_sexe}{1- Masculin 2- Feminin 3- Indéterminé}


}
\source{DEMERSTEM EU project}
\keyword{datasets}
